test(mock-agent): replace indieauth agent
